add collectiveId and a hook to fetch user by stored fid that can be used by all components
